Type : 기존에 있던 타입 폴더 제거 후 도메인별 타입 분리
