+++ src/types/page.ts
@@ -1,0 +1,6 @@
+export interface PageInfo {
+  page: number;
+  size: number;
+  totalElements: number;
+  totalPages: number;
+}
