+++ src/types/question.ts
@@ -1,0 +1,9 @@
+import { PageInfo } from './page';
+import { Post } from './post';
+
+export interface Question {
+  code: number;
+  data: Post[];
+  message: string;
+  pageInfo: PageInfo;
+}
